Move Shibboleth authn parsing to utils
Move the logic that parses the authentication object (created by
Shibboleth) into a separate module under a new utils directory.
